// src/api/back/models/PlayerInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
/**
 *
 * @export
 * @interface PlayerInfo
 */
export interface PlayerInfo {
  /**
   *
   * @type {string}
   * @memberof PlayerInfo
   */
  hero: string;
  /**
   *
   * @type {number}
   * @memberof PlayerInfo
   */
  team: number;
  /**
   *
   * @type {string}
   * @memberof PlayerInfo
   */
  steamId: string;
  /**
   *
   * @type {number}
   * @memberof PlayerInfo
   */
  level: number;
  /**
   *
   * @type {number}
   * @memberof PlayerInfo
   */
  posX: number;
  /**
   *
   * @type {boolean}
   * @memberof PlayerInfo
   */
  bot: boolean;
  /**
   *
   * @type {number}
   * @memberof PlayerInfo
   */
  posY: number;
  /**
   *
   * @type {Array<string>}
   * @memberof PlayerInfo
   */
  items: Array<string>;
  /**
   *
   * @type {number}
   * @memberof PlayerInfo
   */
  kills: number;
  /**
   *
   * @type {number}
   * @memberof PlayerInfo
   */
  deaths: number;
  /**
   *
   * @type {number}
   * @memberof PlayerInfo
   */
  assists: number;
  /**
   *
   * @type {string}
   * @memberof PlayerInfo
   */
  name: string;
}

export function PlayerInfoFromJSON(json: any): PlayerInfo {
  return PlayerInfoFromJSONTyped(json, false);
}

export function PlayerInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlayerInfo {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    hero: json["hero"],
    team: json["team"],
    steamId: json["steam_id"],
    level: json["level"],
    posX: json["pos_x"],
    bot: json["bot"],
    posY: json["pos_y"],
    items: json["items"],
    kills: json["kills"],
    deaths: json["deaths"],
    assists: json["assists"],
    name: json["name"]
  };
}

export function PlayerInfoToJSON(value?: PlayerInfo | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    hero: value.hero,
    team: value.team,
    steam_id: value.steamId,
    level: value.level,
    pos_x: value.posX,
    bot: value.bot,
    pos_y: value.posY,
    items: value.items,
    kills: value.kills,
    deaths: value.deaths,
    assists: value.assists,
    name: value.name
  };
}